Experiments/Bots/RAG_Apps/multi_source_bot/requirements.txt
pypdf
langchain_openai 
langchain_core
python-dotenv
langchain_community
langserve
bs4
faiss-cpu
groq
beautifulsoup4
wikipedia
arxiv
langchainhub
PyPDF2
langchain-objectbox